split to directories
