+++ .gitignore
@@ -56,7 +56,6 @@ dkms.conf
 
 # test files
 main.c
-ft_strtol.c
 out.txt
 libftTester/
 libft-unit-test/
