+++ Makefile
@@ -1,75 +1,69 @@
 NAME = libft.a
-CFLAGS = -Wall -Wextra -Werror
-ARFLAGS = cr
-SRCS = ft_atoi.c \
-		ft_atoi.c \
-		ft_bzero.c \
-		ft_calloc.c \
-		ft_isalnum.c \
-		ft_isalpha.c \
-		ft_isascii.c \
-		ft_isdigit.c \
-		ft_isprint.c \
-		ft_itoa.c \
-		ft_memccpy.c \
-		ft_memchr.c \
-		ft_memcmp.c \
-		ft_memcpy.c \
-		ft_memmove.c \
-		ft_memset.c \
-		ft_putchar_fd.c \
-		ft_putendl_fd.c \
-		ft_putnbr_fd.c \
-		ft_putstr_fd.c \
-		ft_split.c \
-		ft_strchr.c \
-		ft_strdup.c \
-		ft_striteri.c \
-		ft_strjoin.c \
-		ft_strlcat.c \
-		ft_strlcpy.c \
-		ft_strlen.c \
-		ft_strmapi.c \
-		ft_strncmp.c \
-		ft_strnstr.c \
-		ft_strrchr.c \
-		ft_strtrim.c \
-		ft_substr.c \
-		ft_tolower.c \
-		ft_toupper.c
+SRCS = $(addprefix memory/,\
+			ft_bzero.c \
+			ft_calloc.c \
+			ft_memccpy.c \
+			ft_memchr.c \
+			ft_memcmp.c \
+			ft_memcpy.c \
+			ft_memmove.c \
+			ft_memset.c	) \
+		$(addprefix output/, \
+			ft_putchar_fd.c \
+			ft_putendl_fd.c \
+			ft_putnbr_fd.c \
+			ft_putstr_fd.c) \
+		$(addprefix type/, \
+			ft_isalnum.c \
+			ft_isalpha.c \
+			ft_isascii.c \
+			ft_isdigit.c \
+			ft_isprint.c \
+			ft_tolower.c \
+			ft_toupper.c) \
+		$(addprefix string/, \
+			ft_atoi.c \
+			ft_itoa.c \
+			ft_split.c \
+			ft_strchr.c \
+			ft_strdup.c \
+			ft_striteri.c \
+			ft_strjoin.c \
+			ft_strlcat.c \
+			ft_strlcpy.c \
+			ft_strlen.c \
+			ft_strmapi.c \
+			ft_strncmp.c \
+			ft_strnstr.c \
+			ft_strrchr.c \
+			ft_strtol.c \
+			ft_strtrim.c) \
+		$(addprefix list/, \
+			ft_lstadd_back.c \
+			ft_lstadd_front.c \
+			ft_lstclear.c \
+			ft_lstdelone.c \
+			ft_lstiter.c \
+			ft_lstlast.c \
+			ft_lstmap.c \
+			ft_lstnew.c \
+			ft_lstsize.c)
 OBJS = $(SRCS:.c=.o)
 
-BONUS_SRCS = ft_lstadd_back.c \
-				ft_lstadd_front.c \
-				ft_lstclear.c \
-				ft_lstdelone.c \
-				ft_lstiter.c \
-				ft_lstlast.c \
-				ft_lstmap.c \
-				ft_lstnew.c \
-				ft_lstsize.c
-BONUS_OBJS = $(BONUS_SRCS:.c=.o)
-
-ifdef BONUS
-	LIBFT_OBJS = $(OBJS) $(BONUS_OBJS)
-else
-	LIBFT_OBJS = $(OBJS)
-endif
+CC = gcc
+CFLAGS = -Wall -Wextra -Werror -I .
+ARFLAGS = cr
 
 all: $(NAME)
 
-$(NAME): $(LIBFT_OBJS)
+$(NAME): $(OBJS)
 	$(AR) $(ARFLAGS) $@ $^
-
-bonus: $(OBJS) $(BONUS_OBJS)
-	make BONUS=42
 
 %.o: %.c
 	$(CC) $(CFLAGS) -c $< -o $@
 
 clean:
 	$(RM) $(OBJS)
-	$(RM) $(BONUS_OBJS)
 
 fclean: clean
 	$(RM) $(NAME)
