+++ memory/ft_calloc.c
@@ -1,30 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstadd_back.c                                   :+:      :+:    :+:   */
+/*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/04 19:22:37 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/22 20:03:13 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/03 01:17:23 by reasuke           #+#    #+#             */
+/*   Updated: 2023/10/12 19:06:41 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_lstadd_back(t_list **lst, t_list *new)
+void	*ft_calloc(size_t count, size_t size)
 {
-	t_list	*cur;
+	void	*ret;
 
-	if (!lst)
-		return ;
-	if (!*lst)
-	{
-		*lst = new;
-		return ;
-	}
-	cur = *lst;
-	while (cur->next)
-		cur = cur->next;
-	cur->next = new;
+	if (count && SIZE_MAX / count < size)
+		return (NULL);
+	ret = malloc(count * size);
+	if (!ret)
+		return (NULL);
+	return (ft_memset(ret, 0, count * size));
 }
