+++ memory/ft_memchr.c
