+++ string/ft_striteri.c
