+++ string/ft_strmapi.c
