+++ string/ft_strncmp.c
