+++ string/ft_strnstr.c
