+++ string/ft_strtrim.c
