+++ type/ft_isascii.c
